requirements.txt
torch>=2.3.0
numpy>=1.24
python-chess>=1.999
tqdm>=4.66
PyYAML>=6.0 